test: add unit test coverage for core generic committer v3
